Renaming RSVPStore (2/2)
